green case 3 - all

+++ src/transactions/deleteEmployee.ts
@@ -1,10 +1,15 @@
 import { Transaction } from "./Transactions";
 import { EmployeeRepository } from "../repositories";
+import { TransactionFormatError } from "../errors/TransactionFormatError";
 
 export function buildDeleteEmployeeTransaction(
     employeeRepository: EmployeeRepository
 ): Transaction {
     return async function(employeeId: string): Promise<void> {
+        if (!employeeId) {
+            throw new TransactionFormatError("DelEmp");
+        }
+
         await employeeRepository.deleteById(parseInt(employeeId));
     };
 }
